requirements.txt
scrypt==0.9.4
bcrypt==5.0.0
argon2-cffi==25.1.0
psutil==7.1.3
matplotlib==3.10.7
scipy==1.16.3